Move EndTime and StartTime to lib directory Deal with requires and dependency changes
