+++ app/controllers/events_controller.rb
@@ -1,3 +1,6 @@
+require 'start_time'
+require 'end_time'
+
 class EventsController < ApplicationController
   #require 'delorean'
