+++ lib/end_time.rb
