+++ lib/start_time.rb
